+++ spec/unit/installer/file_references_installer_spec.rb
@@ -8,6 +8,12 @@ module Pod
       @project = Project.new(config.sandbox.project_path)
       @project.add_pod_group('BananaLib', fixture('banana-lib'))
       @installer = Installer::FileReferencesInstaller.new(config.sandbox, [@pod_target], @project)
+
+      @pod_target_no_header = fixture_pod_target('BananaNoHeaderLib.podspec')
+      @file_accessor_no_header = @pod_target_no_header.file_accessors.first
+      @project_no_header = Project.new(config.sandbox.project_path)
+      @project_no_header.add_pod_group('BananaLib', fixture('banana-lib'))
+      @installer_no_header = Installer::FileReferencesInstaller.new(config.sandbox, [@pod_target_no_header], @project_no_header)
     end
 
     #-------------------------------------------------------------------------#
@@ -153,6 +159,13 @@ module Pod
           mappings.should == {
             (headers_sandbox + 'Bananalib') => [header],
             (headers_sandbox + 'Bananalib/SubDir') => [header_subdir],
+          }
+        end
+
+        it 'vendored frameworks without header' do
+          headers_sandbox = Pathname.new('BananaNoHeaderLib')
+          mappings = @installer_no_header.send(:vendored_frameworks_header_mappings, headers_sandbox, @file_accessor_no_header)
+          mappings.should == {
           }
         end
       end